Refactoring + DEBUG_ONLY mode
+ Add DEBUG_ONLY mode
+ Move logger::error to separate file
+ Move logger::ProcessVars and logger::StrToLen to separate file
+ Refactor descriptions
+ Refactor spaces and tabs
